+++ archon/actor/delegate.py
@@ -97,6 +97,8 @@ class ExposureDelegate(Generic[Actor_co]):
 
         self.expose_data: ExposeData | None = None
 
+        self.last_exposure_no: int = -1
+
         self.use_shutter: bool = True
         self.is_writing: bool = False
 
@@ -350,6 +352,10 @@ class ExposureDelegate(Generic[Actor_co]):
         except Exception as err:
             return self.fail(f"Failed reading out: {err}")
 
+        if len(c_fdata) == 0:
+            self.command.error("No data was fetched.")
+            return False
+
         self.command.debug(f"Readout completed in {time()-t0:.1f} seconds.")
 
         if write is False:
@@ -388,6 +394,8 @@ class ExposureDelegate(Generic[Actor_co]):
             )
             for fd in fdata
         ]
+
+        self.last_exposure_no = fdata[0]["exposure_no"]
 
         # Prepare checksum information.
         write_checksum: bool = self.config["checksum.write"]